requirements.txt
torch
torchvision
timm
PyYAML
tqdm
einops
scikit-learn